Make argparse separate known and unknown options that occur in a group together.
For example, argparse --ignore-unknown h -- -ho will now set set $argv to -o and $argv_opts to -h (i.e. -ho is split into -h and -o).
Previously, it would set $argv to -ho, and $argv_opts to empty.
With this change, the "Limitations" section of argparse's man page has been removed, and the examples merged into the description of the
-i/--ignore-unknown option. (Note: there was another 'limitation' mentioned in the 'limitations' section: that everything occuring after an unknown option
in a group was considered an argument to an option; the documentation has been reworded to make it clear that this is intended behaviour, as unknown options
are always treated as taking optional arguments, and modifying that behaviour would be a breaking change and not a bug fix).