Added support to fish_opt for defining a long flag with no short flag.
Specifically, this now makes the -s/--short option to fish_opt optional when the -l/--long option is given.
This commit does not modify argparse, as it already supports defining long flags without a corresponding short flag,
however fish_opt would never take advantage of this feature.

Note that due to a limitation in argparse, fish_opt will give an error if you try to define a one-character --long flag without
also providing a --short option.

For backwards compatibility, the --long-only flag is still included with fish_opt,
and when used with -s/--short, will behave as before (the short flag is still defined, but argparse will fail if it is actually used by the parsed arguments,
moreover the _flag_ option variables will not be defined). This can however be used to define a one character long flag.