Update pinned stable Rust to 1.89
Rust 1.89 introduces new lifetime-related lints:
https://github.com/rust-lang/rust/pull/138677

The current code results in warnings about lifetime syntax mismatches,
e.g. in `fn kind(&self) -> Kind`, where the lifetime is elided in `&self` but
hidden in `Kind`. We might want to address these issues in the code instead of
disabling the warning. To prevent issues for unrelated development, the warning
is enabled for now.