Stop pinning stable Rust
Having stable Rust pinned requires updating the pinned version whenever a new
version of stable Rust is released (every 6 weeks, ignoring patch version).
This requires manual effort without providing much benefit.

In the current configuration, running `build_tools/check.sh` fails when using
the stable toolchain when either the local stable toolchain or the pinned
version in `.github/actions/rust-toolchain@stable/action.yml` are outdated.
The former is a nice reminder to update the installed stable toolchain.
Alternatively, one can use the toolchain of a specific version if the up-to-date
stable toolchain is undesirable.
The latter can be very annoying. It results in automatic failure to run the
`check.sh` script at least once every 6 weeks, which needs to be resolved by
updating the pinned version (ideally accompanied by addressing all new lints).
Developers not familiar with the pinning setup might get blocked and annoyed by
such failures, and for all developers it is required to
- stop using the `check.sh` script
- or modify it to eliminate the check
- or update the pinned version
- or stop development entirely until the issue is resolved (temporarily if the
  current setup is kept)

Between stable Rust updating and the pinned version update being merged into
master, the development workflow is hindered and might result in pointless
rebasing, modifications of the `check.sh` script, or other annoyances.

Not pinning stable Rust eliminates all these issues.
The advantage of pinning is that changes in the new version do not automatically
break CI runs which ran fine on older versions of Rust. I think this is mostly
an issue for code not affected by the PR for which CI runs. Regularly running
checks on the master branch using the beta toolchain and fixing detected issues
before they make it to the stable toolchain eliminates the problem.