Add tooling for running ShellCheck
ShellCheck needs to be told which files to check. Identifying these files
automatically is nontrivial. File extensions are insufficient, since we have
several shell scripts which do not have an extension. Instead, we look for
relevant shebangs in the first line of all files in the repo and consider these
files for checking.

Because there might be files which should not be checked for whatever reason, we
add a way to exclude them by putting their paths into a dedicated file. At the
moment, no files are excluded.

Add a script for generating the resulting list of paths to check. The script can
also be used to check if the list is up-to-date.