Add xtask setup for running cargo with version set
This is an initial implementation of the cargo xtask approach
https://github.com/matklad/cargo-xtask

The alias in `.cargo/config.toml` allows running xtasks as
`cargo xtask args for xtask`
which would result in the xtask code being passed the three arguments
`args`, `for`, and `xtask`. (The first argument it gets is its own
binary path, but we ignore that.)

For now, the only use of this is to ensure that the `FISH_BUILD_VERSION`
env var is set to the appropriate value when running cargo commands.
To utilize the feature, pass `cargo` or `c` as the first argument of the
xtask invocation, and add arguments which should be forwarded to cargo.
E.g. `cargo xtask c build` would result in running `cargo build` with
`FISH_BUILD_VERSION` set.