Add xtask for building HTML docs
This task is a bit annoying to implement because `sphinx-build` depends
on `fish_indent`, so that needs to be built, and the binary location
added to `PATH`.

Then, we also need to set up a directory for `sphinx-build` to put its
output. This will be `target/fish-html` for non-CMake builds.