Prohibit duplicate variable names with function -a and -V options
It is now an error to use the same variable name twice with
--inherit-variable and/pr --argument-names.
The previous behaviour was:
    1. duplicate --inherit-variable's worked as if you only inherited
    the variable once
    2. Later --argument-names override earlier ones
    3. An --inherit-variable overrides an --argument-names
Thus this change is backwards incompatible, however, such duplicate
names are not useful and where likely unintentional, and none of the
test cases were depending on the old behaviour.

(I do note that there was code written to specifically remove any
duplicate --inherit-variable's, but the only difference this made was
that the duplicates wouldn't be displayed by the functions or type
builtin. This seems to have been done to preserve the prior behaviour
when it was a HashMap.)