Always rebuild if `FISH_BUILD_VERSION` is not set
This is an important step towards ensuring that the version info which
is embedded into build artifacts is accurate. If `FISH_BUILD_VERSION` is
not set at build time, the version string will by obtained at build
time, and cargo will unconditionally rebuild on every command until
`FISH_BUILD_VERSION` is set; then it will only rebuild if it changes.

`cargo xtask cargo` should be used instead of plain `cargo` to
automatically set `FISH_BUILD_VERSION`.

Rust-analyzer will also always rebuild by default now, but there it
should not make much of a difference, since when rust-analyzer decides
to rebuild, a rebuild is necessary regardless of version changes.
The version is unimportant for rust-analyzer builds, so to avoid some
work, it can be configured to run with `FISH_BUILD_VERSION` set, as
described in the comment included in this commit.