Skip sending OSC7 sequences when running inside SSH session
Fixes an issue that affects at least Apple Terminal when connecting to a remote server that runs fish over ssh. Fish emits an OSC7 sequence signaling the current working directory in the server and as the path is not valid on the client a question mark is shown over the icon in the Terminal title bar.

Fixes #11777.