Do not require the variadic --argument-names to be the last one
This commit makes it so that an argument-name ending with a ...
in a function's --argument-names list is not required to be the last.
There can still only be one such argument name.
The idea is that a '...' argument name will consume as many arguments
from argv as possible, whilst still leaving one argument for each normal
argument name (if possible).