config_paths: don't fall back to compiled-in doc dir for in-tree CMake builds
Commit b758c0c335a (Relax the requirement that we find a working 'doc'
directory in order for fish to be relocatable., 2014-01-17)
came up with a smart fallback for $__fish_help_dir.

Commit c2a8de48735 (Make fish find config directories in source tree,
2016-09-23) made in-tree builds like "cmake . && ninja && ./fish"
use the files from the repository instead of compiled-in paths.

I don't think anyone uses in-tree builds, and even if, they should
probably not be using /usr/share/doc/fish (other not-installed variants
like "target/debug/fish" and Cmake's "build/fish" don't either).
Let's restrict this special behavior to relocatable trees where
it was explicitly added (the fallback doesn't seem great but it's
mostly harmless).