Added ability to exclude parts of code from fish_indent reformatting.
This makes fish_indent (when not in --only-indent or
--only-unindent mode) recognise the following special comments:
    #!fish_indent: on
    #!fish_indent: off
Where there can be any amount of whitespace around "on" and "off".
Any text starting from a "#!fish_indent: off", and continuing to the
start of the next "#!fish_indent: on" (or end of file), is output
verbatim by fish_indent, and not reformatted.

The implementation of this prioritises ease of maintainance vs.
performance: thus the text in a "#!fish_indent: off" is still formated
like normal, but this formated text is then replaced with the original
text when the "#!fish_indent: on" (or end of file) is encountered.
Thus nothing special needs to be done if the formatting code is changed,
provided that:
    1. a "#!fish_indent" directive is always proccessed before any text
    after it.
    2. the formatted-text before a "#!fish_indent: on" directive is not
    modified once said directive has been proccessed.

The downside of this approach is that it will waste time and memory
generating formatted code that is then simply discarded.

Only a simple test case is included, demonstrating that this feature
works within a single command line. A later commit however will
efectively make build_tools/check.fish itself a test case for this
feature.