Feature flag for turning off TTY querying
Experience with OSC 133 and kitty keyboard protocol enabling sequences
have shown that a lot of users are still on incompatible terminals.
It's not always easy to fix those terminals straight away.  There are
probably some environments where primary device attribute queries
are not answered.

Add a feature flag (joining keyboard-protocols and mark-prompt)
to allow users to turn this off.

When the terminal fails to respond to primary device attribute, we
already print an error pointing to "help terminal-compatibility".
In the "primary device attribute", point out the feature flag.