printf: add support for explicit arg positions
This allows specifying explicit argument positions in the format string
of the `printf` command, e.g. `'%2$*1$.*2$d' 10 4`

Note that this feature is incompatible with applying arguments to the
format string until all arguments are used up.
Also note that using explicit argument positions is all-or-nothing.
If they appear in the format string, referring to arguments implicitly
is not possible.

Documentation updates and tests are missing for now.